+++ src/core/cast.hpp
