+++ src/plugins/implementation/plugin_loader.cpp
@@ -26,7 +26,7 @@
 
 #include <core/ilogger.hpp>
 #include <core/ilogger_factory.hpp>
-#include <core/down_cast.hpp>
+#include <core/cast.hpp>
 #include <system/filesystem.hpp>
 #include <database/idatabase_plugin.hpp>
